Adding Some Projects to GIT
This project replaces another project because I had some problems with
the previous version of eclipse I've entirely remade the workspace. Also
because I needed re-factoring my code which was generating an
application but the previous generated code was missing behavior. So the
structure of this second big version is a little bit different. Actually
it generates less code (because the transition from the last version is
a little bit slow). The current version don't generates a running
application like the previous one:
https://www.youtube.com/watch?v=tT2ub59OjA4 But instead the focus is in
generating a Modular Application (which is being a little difficult due
to dependences affecting database synchronization: each module must be
able to survive by itself). This project also uses a little JavaScript
code generator (which is not included yet) the JavaScript code is able
to sort the tables according to the dependences and then generate a
synchronization code in java language. I'll add it in another commit