No significant changes from the viewpoint of the generated code
Adding 'begin_reserve()' and 'end_reserve()' to basic_structure.mtl

using begin_reserve() and end_reserve() in build_android_core_manifest.
but It's clearly showing no results in to the generated code

Adding AndroidManifest.xml (empty) to generateElement() in
'AndroidSqliteDatabaseSingle.mtl'

Adding AVOID-TODO.txt file. It's a warning saying that you should prefer
to avoid Replacing the Type of an uml::Property. Because it causes
eclipse to froze. Instead it would be better to Delete the old property
and create a new one with the same name. then set the type