Adding new tests to the project com.uisleandro.test.nullproperties
Now It's testing for repeated property names (which could cause problems
during the code generation)

The model has again being changed. The property name was part of 'user'
and 'reseller' which inherits from user.

A lot of model corrections needed to be made in order to make the new
test pass. But It's a problem because the transformation acaually causes
these problems, for example when a type of some 'Property' is changed..
so It will generate a new one (not changing the type of the existing
one) It produces two properties with the same name. Wich impacts really
nad in the code genration.